Refactor class-based example to facilitate adding more types of examples.
